@@ -2,31 +2,18 @@
 return [
     'simplefm' => [
         'connection' => [
-            'uri' => 'https://example.com',
+            'base_uri' => 'https://example.com',
             'username' => 'filemaker-username',
             'password' => 'filemaker-password',
             'database' => 'sample-database',
 
-            // The following setting is required and you must set an appropriate HTTP client in your container.
-            'http_client' => 'http-client.container.key',
-
-            // The following settings are optional and can be ommited. These are keys pointing to objects in your
-            // dependency container. This package provides ready-made factories which you can use.
-            'identity_handler' => 'soliant.simplefm.expressive.identity-handler',
-            'logger' => 'soliant.simplefm.expressive.logger',
-        ],
-
-        'result_set_client' => [
             // Set this to the time zone of your file maker server
             'time_zone' => 'America/Los_Angeles',
         ],
 
-        'authenticator' => [
-            'identity_layout' => 'identity-layout',
-            'username_field' => 'username-field',
-
-            // The following setting is optional.
-            'identity_handler' => 'soliant.simplefm.expressive.identity-handler',
+        'client' => [
+            // The following setting is required and you must set an appropriate HTTP client in your container.
+            'http_client' => 'http-client.container.key',
         ],
 
         'repository_builder' => [
@@ -37,16 +24,6 @@ return [
             'additional_types' => [
                 'type-name-in-xml' => 'container-key',
             ],
-        ],
-
-        // This config is only required when you use "soliant.simplefm.expressive.identity-handler"
-        'identity_handler' => [
-            'key' => 'some-strong-encryption-key',
-        ],
-
-        // This config is only required when you use "soliant.simplefm.expressive.logger"
-        'logger' => [
-            'path' => '/path/to/simplefm/log',
         ],
     ],
 ];
